+++ custom_datasets/__init__.py
@@ -1,1 +1,1 @@
-from custom_datasets.pastis import PastisDataset
+from custom_datasets.flood_dataset import FloodDataset
